캐릭터 모듈화
